requirements.txt
opencv-python
numpy
Pillow
requests
urllib3
